Выполнение 1 пункта ТЗ.
Унифицирована работа с нижним навигационным меню для обеспечения его стабильности работы. Целый день убил на то, что пытался разобраться почему у меня в одном фрагменте бургер-кнопка вызывается, а в другом фрагменте не вызывается. А после перезапуска активити в результате работы фрагмента с настройками вообще перестаёт работать бургер-кнопка во всех фрагментах.
Теперь нижнее навигационное меню корректно работает, поскольку создано единожды в "MainActivity".
Лишний код из фрагментов убран.

Прошу ещё сутки для доработки 3 ДЗ.
Приступаю к созданию других фрагментов и обеспечению навигации между ними.

Ссылка на сайт с репозиторием проекта:
https://github.com/AVKhakhalin/NasaApplication
