+++ app/src/main/java/com/example/nasaapplication/controller/navigation/contents/ViewPagerAdapter.kt
@@ -4,6 +4,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentStatePagerAdapter
 import com.example.nasaapplication.controller.ConstantsController
+import com.example.nasaapplication.ui.activities.MainActivity
 import com.example.nasaapplication.ui.fragments.contents.DayPhotoFragment
 import com.example.nasaapplication.ui.fragments.contents.SearchWikiFragment
 
@@ -18,7 +19,7 @@ class ViewPagerAdapter(private val fragmentManager: FragmentManager):FragmentSta
         return when(position){
             0 -> fragments[ConstantsController.DAY_PHOTO_FRAGMENT]
             1 -> fragments[ConstantsController.SEARCH_WIKI_FRAGMENT]
-            else ->fragments[ConstantsController.DAY_PHOTO_FRAGMENT]
+            else -> fragments[ConstantsController.DAY_PHOTO_FRAGMENT]
         }
     }
