+++ app/src/main/java/com/example/nasaapplication/ui/activities/MainActivity.kt
@@ -1,12 +1,17 @@
 package com.example.nasaapplication.ui.activities
 
+import android.content.Context
 import android.content.SharedPreferences
 import android.os.Build
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.PersistableBundle
-import android.view.View
+import android.util.AttributeSet
+import android.view.*
 import android.widget.FrameLayout
+import android.widget.Toast
+import androidx.core.content.ContextCompat
+import androidx.fragment.app.Fragment
 import androidx.viewpager.widget.ViewPager
 import com.example.nasaapplication.R
 import com.example.nasaapplication.controller.navigation.contents.NavigationContent
@@ -16,6 +21,8 @@ import com.example.nasaapplication.controller.navigation.dialogs.NavigationDialo
 import com.example.nasaapplication.controller.navigation.dialogs.NavigationDialogsGetter
 import com.example.nasaapplication.databinding.ActivityMainBinding
 import com.example.nasaapplication.ui.ConstantsUi
+import com.example.nasaapplication.ui.fragments.contents.SettingsFragment
+import com.google.android.material.bottomappbar.BottomAppBar
 import com.google.android.material.tabs.TabLayout
 
 class MainActivity: AppCompatActivity(), NavigationDialogsGetter, NavigationContentGetter {
@@ -27,6 +34,8 @@ class MainActivity: AppCompatActivity(), NavigationDialogsGetter, NavigationCont
     private var isThemeDay: Boolean = true
     // Binding
     lateinit var binding: ActivityMainBinding
+    // Bottom navigation menu
+    private var isMain = false
     //endregion
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -62,6 +71,9 @@ class MainActivity: AppCompatActivity(), NavigationDialogsGetter, NavigationCont
         binding.tabLayout.visibility = View.VISIBLE
         binding.activityFragmentsContainer.visibility = View.INVISIBLE
 
+        // Установка Bottom Navigation Menu
+        setBottomAppBar()
+
         // Отображение содержимого макета
         setContentView(binding.root)
     }
@@ -91,5 +103,82 @@ class MainActivity: AppCompatActivity(), NavigationDialogsGetter, NavigationCont
 
     override fun getNavigationContent(): NavigationContent {
         return navigationContent
+    }
+
+    //region УСТАНОВКА BOTTOM NAVIGATION MENU
+    private fun setBottomAppBar() {
+        this.setSupportActionBar(binding.bottomNavigationMenu.bottomAppBar)
+//        setHasOptionsMenu(true)
+
+        switchBottomAppBar(this)
+        binding.bottomNavigationMenu.bottomAppBarFab.setOnClickListener {
+            switchBottomAppBar(this)
+        }
+    }
+
+    // Переключение режима нижней навигационной кнопки BottomAppBar
+    // с центрального на крайнее левое положение и обратно
+    private fun switchBottomAppBar(context: MainActivity) {
+        if (isMain) {
+            isMain = false
+            binding.bottomNavigationMenu.bottomAppBar.navigationIcon = null
+            binding.bottomNavigationMenu.bottomAppBar.fabAlignmentMode = BottomAppBar.FAB_ALIGNMENT_MODE_END
+            binding.bottomNavigationMenu.bottomAppBarFab.setImageDrawable(
+                ContextCompat.getDrawable(
+                    context, R.drawable.ic_back_fab
+                )
+            )
+            binding.bottomNavigationMenu.bottomAppBar.replaceMenu(R.menu.bottom_menu_bottom_bar_other_screen)
+        } else {
+            isMain = true
+            binding.bottomNavigationMenu.bottomAppBar.navigationIcon =
+                ContextCompat.getDrawable(context, R.drawable.ic_hamburger_menu_bottom_bar)
+            binding.bottomNavigationMenu.bottomAppBar.fabAlignmentMode = BottomAppBar.FAB_ALIGNMENT_MODE_CENTER
+            binding.bottomNavigationMenu.bottomAppBarFab.setImageDrawable(
+                ContextCompat.getDrawable(
+                    context, R.drawable.ic_plus_fab
+                )
+            )
+            binding.bottomNavigationMenu.bottomAppBar.replaceMenu(R.menu.bottom_menu_bottom_bar)
+        }
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu): Boolean {
+        menuInflater.inflate(R.menu.bottom_menu_navigation, menu)
+        return true
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when (item.itemId) {
+            R.id.action_bottom_bar_navigation_to_send -> toast("Перейти к отправке")
+            R.id.action_bottom_bar_navigation_to_archive -> toast("Перейти в архив")
+            R.id.action_bottom_bar_navigation_to_observe -> toast("Перейти к картинке дня")
+            R.id.action_bottom_bar_save -> toast("Сохранение")
+            R.id.action_bottom_bar_search -> toast("Поиск")
+            R.id.action_bottom_bar_settings -> navigationContent?.let{
+                this.findViewById<ViewPager>(R.id.view_pager).visibility =
+                    View.INVISIBLE
+                this.findViewById<TabLayout>(R.id.tab_layout).visibility =
+                    View.INVISIBLE
+                this.findViewById<FrameLayout>(R.id.activity_fragments_container)
+                    .visibility = View.VISIBLE
+                it.showSettingsFragment(false)
+            }
+            android.R.id.home -> {
+                navigationDialogs?.let {
+                    it.showBottomNavigationDrawerDialogFragment(this)
+                }
+            }
+        }
+        return super.onOptionsItemSelected(item)
+    }
+    //endregion
+
+    // Метод для отображения сообщения в виде Toast
+    private fun MainActivity.toast(string: String?) {
+        Toast.makeText(this, string, Toast.LENGTH_SHORT).apply {
+            setGravity(Gravity.BOTTOM, 0, 250)
+            show()
+        }
     }
 }
