+++ app/src/main/java/com/example/nasaapplication/ui/fragments/contents/DayPhotoFragment.kt
@@ -1,18 +1,14 @@
 package com.example.nasaapplication.ui.fragments.contents
 
 import android.content.Context
-import android.content.DialogInterface
 import android.os.Bundle
 import android.view.*
-import android.widget.FrameLayout
 import android.widget.TextView
 import android.widget.Toast
 import androidx.constraintlayout.widget.ConstraintLayout
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
-import androidx.viewpager.widget.ViewPager
 import coil.load
 import com.example.nasaapplication.R
 import com.example.nasaapplication.controller.navigation.contents.NavigationContent
@@ -23,10 +19,8 @@ import com.example.nasaapplication.databinding.FragmentDayPhotoBinding
 import com.example.nasaapplication.ui.ConstantsUi
 import com.example.nasaapplication.ui.activities.MainActivity
 import com.example.nasaapplication.ui.utils.ViewBindingFragment
-import com.google.android.material.bottomappbar.BottomAppBar
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.chip.Chip
-import com.google.android.material.tabs.TabLayout
 import java.util.*
 
 class DayPhotoFragment: ViewBindingFragment<FragmentDayPhotoBinding>(FragmentDayPhotoBinding::inflate) {
@@ -84,7 +78,6 @@ class DayPhotoFragment: ViewBindingFragment<FragmentDayPhotoBinding>(FragmentDay
                     binding.pODImageView.load(url) {
                         lifecycle(this@DayPhotoFragment)
                         error(R.drawable.ic_load_error_vector)
-//                        placeholder(R.drawable.ic_downloading)
                     }
                     // Показать описание фотографии дня
                     bottomSheetDescriptionTitle.text = serverResponseData.title
@@ -125,9 +118,6 @@ class DayPhotoFragment: ViewBindingFragment<FragmentDayPhotoBinding>(FragmentDay
         binding.fragmentDayPhotoCurrentDateTextView.text =
             "${ConstantsUi.DAY_PHOTO_TEXT} ${getDate(0)}"
 
-        // Установка BOTTOM NAVIGATION MENU
-        setBottomAppBar(view)
-
         // Инициализация текстового блока для отображения текущей даты
         currentDateTextView = binding.fragmentDayPhotoCurrentDateTextView
 
@@ -167,72 +157,6 @@ class DayPhotoFragment: ViewBindingFragment<FragmentDayPhotoBinding>(FragmentDay
     private fun setBottomSheetBehavior(bottomSheet: ConstraintLayout) {
         bottomSheetBehavior = BottomSheetBehavior.from(bottomSheet)
         bottomSheetBehavior.state = BottomSheetBehavior.STATE_COLLAPSED
-    }
-    //endregion
-
-    //region МЕТОДЫ ДЛЯ РАБОТЫ С BOTTOM NAVIGATION MENU
-    private fun setBottomAppBar(view: View) {
-        val context = activity as MainActivity
-        context.setSupportActionBar(binding.bottomNavigationMenu.bottomAppBar)
-        setHasOptionsMenu(true)
-
-        binding.bottomNavigationMenu.bottomAppBarFab.setOnClickListener {
-            switchBottomAppBar(context)
-        }
-    }
-
-    // Переключение режима нижней навигационной кнопки BottomAppBar
-    // с центрального на крайнее левое положение и обратно
-    private fun switchBottomAppBar(context: MainActivity) {
-        if (isMain) {
-            isMain = false
-            binding.bottomNavigationMenu.bottomAppBar.navigationIcon = null
-            binding.bottomNavigationMenu.bottomAppBar.fabAlignmentMode = BottomAppBar.FAB_ALIGNMENT_MODE_END
-            binding.bottomNavigationMenu.bottomAppBarFab.setImageDrawable(
-                ContextCompat.getDrawable(
-                    context, R.drawable.ic_back_fab
-                )
-            )
-            binding.bottomNavigationMenu.bottomAppBar.replaceMenu(R.menu.bottom_menu_bottom_bar_other_screen)
-        } else {
-            isMain = true
-            binding.bottomNavigationMenu.bottomAppBar.navigationIcon =
-                ContextCompat.getDrawable(context, R.drawable.ic_hamburger_menu_bottom_bar)
-            binding.bottomNavigationMenu.bottomAppBar.fabAlignmentMode = BottomAppBar.FAB_ALIGNMENT_MODE_CENTER
-            binding.bottomNavigationMenu.bottomAppBarFab.setImageDrawable(
-                ContextCompat.getDrawable(
-                    context, R.drawable.ic_plus_fab
-                )
-            )
-            binding.bottomNavigationMenu.bottomAppBar.replaceMenu(R.menu.bottom_menu_bottom_bar)
-        }
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        super.onCreateOptionsMenu(menu, inflater)
-//        inflater.inflate(R.menu.bottom_menu_bottom_bar, menu)
-    }
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when (item.itemId) {
-            R.id.app_bar_save -> toast("Сохранение")
-            R.id.app_bar_settings -> navigationContent?.let{
-                requireActivity().findViewById<ViewPager>(R.id.view_pager).visibility =
-                    View.INVISIBLE
-                requireActivity().findViewById<TabLayout>(R.id.tab_layout).visibility =
-                    View.INVISIBLE
-                requireActivity().findViewById<FrameLayout>(R.id.activity_fragments_container)
-                    .visibility = View.VISIBLE
-                it.showSettingsFragment(false)
-            }
-            R.id.app_bar_search -> toast("Поиск")
-            android.R.id.home -> {
-                toast("DayPhotoFragment Бургер кнопка")
-                navigationDialogs?.let {
-                    it.showBottomNavigationDrawerDialogFragment(requireActivity())
-                }
-            }
-        }
-        return super.onOptionsItemSelected(item)
     }
     //endregion
