+++ app/src/main/res/drawable/ic_archive.xml
@@ -7,21 +7,21 @@
         android:fillColor="@android:color/transparent"
         android:pathData="M21,8l0,13l-18,0l0,-13"
         android:strokeWidth="2"
-        android:strokeColor="#000000"
+        android:strokeColor="#FFFFFF"
         android:strokeLineCap="round"
         android:strokeLineJoin="round" />
     <path
         android:fillColor="@android:color/transparent"
         android:pathData="M1,3h22v5h-22z"
         android:strokeWidth="2"
-        android:strokeColor="#000000"
+        android:strokeColor="#FFFFFF"
         android:strokeLineCap="round"
         android:strokeLineJoin="round" />
     <path
         android:fillColor="@android:color/transparent"
         android:pathData="M10,12L14,12"
         android:strokeWidth="2"
-        android:strokeColor="#000000"
+        android:strokeColor="#FFFFFF"
         android:strokeLineCap="round"
         android:strokeLineJoin="round" />
 </vector>
