+++ app/src/main/res/drawable/ic_back_fab.xml
@@ -7,14 +7,14 @@
         android:fillColor="@android:color/transparent"
         android:pathData="M9,14l-5,-5l5,-5"
         android:strokeWidth="2"
-        android:strokeColor="#000000"
+        android:strokeColor="#FFFFFF"
         android:strokeLineCap="round"
         android:strokeLineJoin="round" />
     <path
         android:fillColor="@android:color/transparent"
         android:pathData="M20,20v-7a4,4 0,0 0,-4 -4H4"
         android:strokeWidth="2"
-        android:strokeColor="#000000"
+        android:strokeColor="#FFFFFF"
         android:strokeLineCap="round"
         android:strokeLineJoin="round" />
 </vector>
