+++ app/src/main/res/drawable/ic_send.xml
@@ -7,14 +7,14 @@
         android:fillColor="@android:color/transparent"
         android:pathData="M22,2L11,13"
         android:strokeWidth="2"
-        android:strokeColor="#000000"
+        android:strokeColor="#FFFFFF"
         android:strokeLineCap="round"
         android:strokeLineJoin="round" />
     <path
         android:fillColor="@android:color/transparent"
         android:pathData="M22,2l-7,20l-4,-9l-9,-4l20,-7z"
         android:strokeWidth="2"
-        android:strokeColor="#000000"
+        android:strokeColor="#FFFFFF"
         android:strokeLineCap="round"
         android:strokeLineJoin="round" />
 </vector>
