+++ app/src/main/res/drawable/ic_wikipedia.xml
@@ -4,6 +4,6 @@
     android:viewportWidth="100"
     android:viewportHeight="100">
   <path
-      android:fillColor="#FF000000"
+      android:fillColor="#FFFFFF"
       android:pathData="M94.938,19H76.331l-1.332,0.229v1.788c0,0.21 0.636,0.386 0.812,0.517c0.168,0.142 0.623,0.219 0.811,0.219l1.686,0.084c1.912,0.09 2.494,0.505 3.239,1.205c0.731,0.725 0.851,1.955 0.281,3.684L65.383,69.654l-0.628,-0.187L54.309,45.97c0.151,-0.353 0.242,-0.551 0.242,-0.551l8.553,-17.578c0.994,-1.854 1.957,-3.511 2.446,-4.283c0.894,-1.408 1.328,-1.674 3.763,-1.797c0.498,0 0.688,-0.244 0.688,-0.738v-1.778l-0.125,-0.191c0,0 -10.224,-0.035 -14.879,0l0.002,0.183v1.779c0,0.217 -0.119,0.39 0.046,0.521c0.176,0.142 0.25,0.215 0.439,0.215l0.697,0.089c1.902,0.09 2.754,0.636 3.012,0.979c0.455,0.615 0.668,1.271 -0.434,3.911l-6.585,14.42l-5.95,-13.376c-1.952,-4.274 -2.434,-5.597 0.507,-5.844l0.79,-0.183c0.566,0 0.478,-0.224 0.478,-0.649v-1.87l-0.059,-0.153c0,0 -10.899,0 -15.873,0.003L32,19.262v1.875c0,0.421 0.416,0.529 1.186,0.645c2.754,0.396 2.707,0.713 5.428,6.646c0.407,0.889 1.61,3.57 1.61,3.57l6.203,13.391c0,0 0.704,1.72 1.805,4.39l-9.016,19.763l-0.514,-0.146c0,0 -13.992,-32.418 -18.037,-43.266c-0.427,-1.117 -0.611,-1.953 -0.611,-2.468c0,-1.102 0.904,-1.699 2.712,-1.779l2.751,-0.099c0.563,0 1.483,-0.236 1.483,-0.727v-1.779l-0.829,-0.184c0,0 -17.139,-0.021 -20.602,0L5,19.279v1.872c0,0.336 0.449,0.545 1.228,0.634c2.128,0.129 3.559,0.499 4.229,1.123c0.668,0.628 1.393,2.215 2.333,4.688c5.086,13.424 15.901,37.38 21.166,51.026c1.512,3.724 3.423,4.303 5.732,-0.117c2.379,-4.848 7.179,-15.608 10.68,-23.552c3.275,7.924 7.734,18.645 9.885,23.501c1.712,3.887 3.664,4.271 5.646,0.116c5.183,-12.654 20.319,-50.736 20.319,-50.736c0.654,-1.869 1.568,-3.309 2.746,-4.355c1.169,-1.033 2.909,-1.617 5.349,-1.731c0.494,0 0.688,-0.249 0.688,-0.731v-1.788L94.938,19z"/>
 </vector>
