+++ app/src/main/res/layout/fragment_day_photo.xml
@@ -105,9 +105,5 @@
     <include
         layout="@layout/bottom_sheet_layout" />
 
-    <include
-        android:id="@+id/bottom_navigation_menu"
-        layout="@layout/bottom_navigation_menu_layout" />
-
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
