+++ app/src/main/res/layout/fragment_settings.xml
@@ -52,9 +52,5 @@
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 
-    <include
-        android:id="@+id/bottom_navigation_menu"
-        layout="@layout/bottom_navigation_menu_layout" />
-
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
