+++ app/src/main/res/menu/bottom_menu_bottom_bar.xml
@@ -3,14 +3,14 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <item
-        android:id="@+id/app_bar_save"
+        android:id="@+id/action_bottom_bar_save"
         android:icon="@drawable/ic_save"
         android:title="@string/app_bar_save_title"
         android:fontFamily="@font/convergence"
         app:showAsAction="ifRoom" />
 
     <item
-        android:id="@+id/app_bar_settings"
+        android:id="@+id/action_bottom_bar_settings"
         android:icon="@drawable/ic_settings_menu"
         android:title="@string/app_bar_settings_title"
         android:fontFamily="@font/convergence"
