+++ app/src/main/res/menu/bottom_menu_bottom_bar_other_screen.xml
@@ -3,9 +3,9 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <item
-        android:id="@+id/app_bar_search"
+        android:id="@+id/action_bottom_bar_search"
         android:icon="@drawable/ic_search"
         android:title="@string/app_bar_search_title"
         android:fontFamily="@font/convergence"
-        app:showAsAction="ifRoom" />
+        app:showAsAction="always" />
 </menu>
