+++ app/src/main/res/menu/bottom_menu_navigation.xml
@@ -1,20 +1,24 @@
 <?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
     <group android:checkableBehavior="none">
         <item
-            android:id="@+id/action_navigation_to_archive"
-            android:icon="@drawable/ic_archive"
-            android:title="@string/action_navigation_to_archive_text"
-            android:fontFamily="@font/convergence" />
-        <item
-            android:id="@+id/action_navigation_to_send"
-            android:icon="@drawable/ic_send"
-            android:title="@string/action_navigation_to_send_text"
-            android:fontFamily="@font/convergence" />
-        <item
-            android:id="@+id/action_navigation_to_observe"
+            android:id="@+id/action_bottom_bar_navigation_to_observe"
             android:icon="@drawable/ic_telescope"
             android:title="@string/action_navigation_to_observe_text"
-            android:fontFamily="@font/convergence" />
+            android:fontFamily="@font/convergence"
+            app:showAsAction="always" />
+        <item
+            android:id="@+id/action_bottom_bar_navigation_to_archive"
+            android:icon="@drawable/ic_archive"
+            android:title="@string/action_navigation_to_archive_text"
+            android:fontFamily="@font/convergence"
+            app:showAsAction="always" />
+        <item
+            android:id="@+id/action_bottom_bar_navigation_to_send"
+            android:icon="@drawable/ic_send"
+            android:title="@string/action_navigation_to_send_text"
+            android:fontFamily="@font/convergence"
+            app:showAsAction="always" />
     </group>
 </menu>
